chore(wip): Adds Weather schema, updates existing calendars
