+++ src/settings/settings.service.ts
@@ -16,7 +16,11 @@ import type {
     UndatedCalEvent,
     Version,
 } from "src/@types";
-import { DEFAULT_DATA } from "./settings.constants";
+import {
+    DEFAULT_DATA,
+    DEFAULT_SEASONAL_DATA,
+    DEFAULT_WEATHER_DATA,
+} from "./settings.constants";
 import merge from "deepmerge";
 import { nanoid } from "src/utils/functions";
 import Calendarium from "src/main";
@@ -30,7 +34,6 @@ import {
 } from "./settings.utils";
 import { CHECK, LOADING } from "src/utils/icons";
 import { EventType } from "src/events/event.types";
-import { DEFAULT_SEASONAL_DATA } from "src/schemas/calendar/seasonal";
 
 const SPLITTER = "--- BEGIN DATA ---";
 type CalendarID = string;
@@ -431,7 +434,7 @@ class SettingsServiceClass {
         }
         this.deletedCalendars.push(calendar);
         this.#calendars.delete(calendar.id);
-        
+
         await this.save({ calendar: true });
     }
 
@@ -810,6 +813,11 @@ class SettingsServiceClass {
             }
             if (!("seasonal" in calendar.static)) {
                 (calendar.static as any).seasonal = copy(DEFAULT_SEASONAL_DATA);
+                dirty = true;
+            }
+            if (!("weather" in calendar.static.seasonal)) {
+                (calendar.static as any).seasonal.weather =
+                    copy(DEFAULT_WEATHER_DATA);
                 dirty = true;
             }
         }
